Pass memory to the game from the platform layer Traditional programs do work by asking the Operating System for memory as they need it. This is wrong.
Instead, we allocate *ALL* of the memory we are going to need to run the
entire game until completion up front. If we fail to do this, there is
no point in running the game, because it might fail at any moment and
lose the user's progress or potentially even something worse.

The amount of memory we allocate up front can be tweaked on a
per-platform basis, which could determine the quality of assets that we
load, etc.