Controller input using XInput
We can support controller input now! This is done in such a way that we
have an enum for the states of the buttons on the controller, so you can
figure out if a button was just starting to press or if it is held.
Additionally, we can use the same button state for keyboard keys as
well. And, we use a union between an array of buttons and assigning
names to each key such as `move_left` or `action_up` so that we could
easily have rebindable controls at some point. It's nice to future-proof
your code.